+++ cognite/powerops/clients/data_classes/__init__.py
@@ -96,7 +96,6 @@ PlantApply.model_rebuild()
 PriceAreaApply.model_rebuild()
 RKOMBidApply.model_rebuild()
 RKOMBidCombinationApply.model_rebuild()
-RKOMCombinationBidApply.model_rebuild()
 RKOMProcesApply.model_rebuild()
 ScenarioApply.model_rebuild()
 ScenarioMappingApply.model_rebuild()
